Tightly coupled teams smell refactoring renamed
